requirements.txt
google-adk
google-cloud-bigquery
google-genai
fastmcp
python-dotenv